Calculating Lagrange bases
- Changed type of iterator in `for` loops (int => size_t)
- Fixed assign operator bug

+++ Matrix.cpp
@@ -1,10 +1,10 @@
 #include "Matrix.h"
-int Matrix::size()
+size_t Matrix::size()
 {
 	return this->cols * this->rows;
 }
 
-int Matrix::size() const
+size_t Matrix::size() const
 {
 	return this->cols * this->rows;
 }
